backend/pyproject.toml
[project]
name = "finance-app-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "supabase>=2.4.6",
    "python-dotenv>=1.0.1",
    "anthropic>=0.25.0",
    "google-generativeai>=0.8.0",
    "pydantic>=2.7.1",
]

[tool.uv]
dev-dependencies = []
